+++ app/src/main/java/com/luacevedo/heartbaymax/model/rules/conditions/GreaterThanCondition.java
@@ -2,7 +2,7 @@ package com.luacevedo.heartbaymax.model.rules.conditions;
 
 import com.luacevedo.heartbaymax.model.patient.PatientAttribute;
 
-public class GreaterThanCondition extends BaseCondition<Integer> {
+public class GreaterThanCondition extends BaseCondition<Double> {
 
     private int min;
 
@@ -16,7 +16,7 @@ public class GreaterThanCondition extends BaseCondition<Integer> {
     }
 
     @Override
-    public boolean validate(PatientAttribute<Integer> attribute) {
+    public boolean validate(PatientAttribute<Double> attribute) {
         return attribute.getValue() > min;
     }
 }
